issue #3 feat: MainActivity 수정
MomentwoApp 호출
