+++ app/src/main/java/cord/eoeo/momentwo/ui/MainActivity.kt
@@ -1,0 +1,17 @@
+package cord.eoeo.momentwo.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import cord.eoeo.momentwo.ui.theme.MomentwoTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            MomentwoTheme {
+                MomentwoApp()
+            }
+        }
+    }
+}
